requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.4=py310h4e76f89_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.4.26=hecd8cb5_0
certifi=2022.5.18.1=py310hecd8cb5_0
intel-openmp=2021.4.0=hecd8cb5_3538
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
ncurses=6.3=hca72f7f_2
numexpr=2.8.1=py310hdcd3fac_1
numpy=1.22.3=py310hdcd3fac_0
numpy-base=1.22.3=py310hfd2de13_0
openssl=1.1.1o=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py310he9d5cce_0
pip=21.2.4=py310hecd8cb5_0
pyopengl=3.1.1a1=py310hecd8cb5_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.10.4=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
readline=8.1.2=hca72f7f_1
scipy=1.7.3=py310h3dd3380_0
setuptools=61.2.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=h707629a_0
tk=8.6.11=h3fd3227_1
tzdata=2022a=hda174b7_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_2
